add section separators, change visibility
