+++ crates/rust-client/src/store/sqlite_store/mod.rs
@@ -51,9 +51,6 @@ mod transaction;
 /// Number of sql statements that each connection will cache.
 const SQL_STATEMENT_CACHE_CAPACITY: usize = 32;
 
-// /// How often to run the database maintenance routine.
-// const DATABASE_MAINTENANCE_INTERVAL: Duration = Duration::from_secs(24 * 60 * 60);
-
 // SQLITE STORE
 // ================================================================================================
